Implemented Exception Handling fpr SongsController.

+++ ApiMusic/Controllers/SongsController.cs
@@ -28,7 +28,12 @@ namespace ApiMusic.Controllers
         [HttpGet("{id}")]
         public IActionResult Get(int id)
         {
-            return Ok(_dbContext.Songs.Find(id));
+            var song = _dbContext.Songs.Find(id);
+            if(song == null)
+            {
+                return NotFound("No record found against this Id");
+            }
+            return Ok();
         }
 
         // POST api/<SongsController>
@@ -45,10 +50,17 @@ namespace ApiMusic.Controllers
         public IActionResult Put(int id, [FromBody] Song songObj)
         {
             var song = _dbContext.Songs.Find(id);
-            song.Title = songObj.Title;
-            song.Language = songObj.Language;
-            _dbContext.SaveChanges();
-            return Ok("Record updated successfully");
+            if(song == null)
+            {
+                return NotFound("No record found against this Id");
+            }
+            else
+            {
+                song.Title = songObj.Title;
+                song.Language = songObj.Language;
+                _dbContext.SaveChanges();
+                return Ok("Record updated successfully");
+            }
         }
 
         // DELETE api/<SongsController>/5
@@ -56,9 +68,16 @@ namespace ApiMusic.Controllers
         public IActionResult Delete(int id)
         {
             var song = _dbContext.Songs.Find(id);
-            _dbContext.Songs.Remove(song);
-            _dbContext.SaveChanges();
-            return Ok("Record deleted");
+            if(song == null)
+            {
+                return NotFound("No record found against this Id");
+            }
+            else
+            {
+                _dbContext.Songs.Remove(song);
+                _dbContext.SaveChanges();
+                return Ok("Record deleted");
+            }
         }
     }
 }
